Moved the flush action to db utils
And moved the dbUtils to package 'distributed'
