+++ src/test/java/distributed/TestDatabaseUtils.java
@@ -1,4 +1,4 @@
-package test.java.distributed.workload;
+package test.java.distributed;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
@@ -8,7 +8,7 @@ import java.sql.SQLException;
 import java.sql.Statement;
 
 import main.java.distributed.ConnectionManager;
-import main.java.distributed.workload.DatabaseUtils;
+import main.java.distributed.DatabaseUtils;
 
 import org.junit.After;
 import org.junit.Before;
